ReworkDeteils

+++ Models/PlanDrying.cs
@@ -15,7 +15,7 @@ namespace Test
         public int Id { get; set; }
         public int ValueChamber { get; set; }
         public string MonthDrying { get; set; }
-        public int Utility { get; set; }
+        public decimal Utility { get; set; }
         public int HoursLeftDrying { get; set; }
         public int HoursSpendDrying { get; set; }
         public int LengValue { get; set; }
